add user-based restriction to resource report
